Fix some warnings in kerninfo

+++ src/utils/kerninfo.c
@@ -131,7 +131,7 @@ int main(int argc, char *argv[]) {
 		krnl, dev, CL_KERNEL_WORK_GROUP_SIZE, size_t, &err);
 	ccl_if_err_goto(err, error_handler);
 	g_printf("     Maximum workgroup size                  : %lu\n",
-		k_wg_size);
+		(unsigned long) k_wg_size);
 
 	/* Only show info about CL_KERNEL_PREFERRED_WORK_GROUP_SIZE_MULTIPLE
 	 * if OpenCL version of the underlying platform is >= 1.1. */
@@ -140,26 +140,28 @@ int main(int argc, char *argv[]) {
 			dev, CL_KERNEL_PREFERRED_WORK_GROUP_SIZE_MULTIPLE, size_t, &err);
 		ccl_if_err_goto(err, error_handler);
 		g_printf("     Preferred multiple of workgroup size    : %lu\n",
-			k_pref_wg_size_mult);
+			(unsigned long) k_pref_wg_size_mult);
 	}
 
 	k_compile_wg_size = ccl_kernel_get_workgroup_info_array(krnl, dev,
 		CL_KERNEL_COMPILE_WORK_GROUP_SIZE, size_t*, &err);
 	ccl_if_err_goto(err, error_handler);
 	g_printf("     WG size in __attribute__ qualifier      : (%lu, %lu, %lu)\n",
-		k_compile_wg_size[0], k_compile_wg_size[1], k_compile_wg_size[2]);
+		(unsigned long) k_compile_wg_size[0], 
+		(unsigned long) k_compile_wg_size[1], 
+		(unsigned long) k_compile_wg_size[2]);
 
 	k_loc_mem_size = ccl_kernel_get_workgroup_info_scalar(krnl, dev,
 		CL_KERNEL_LOCAL_MEM_SIZE, cl_ulong, &err);
 	ccl_if_err_goto(err, error_handler);
 	g_printf("     Local memory used by kernel             : %lu bytes\n",
-		k_loc_mem_size);
+		(unsigned long) k_loc_mem_size);
 
 	k_priv_mem_size = ccl_kernel_get_workgroup_info_scalar(krnl, dev,
 		CL_KERNEL_PRIVATE_MEM_SIZE, cl_ulong, &err);
 	ccl_if_err_goto(err, error_handler);
 	g_printf("     Min. private mem. used by each workitem : %lu bytes\n",
-		k_priv_mem_size);
+		(unsigned long) k_priv_mem_size);
 
 	g_printf("\n");
